requirements.txt
aristonremotethermo~=2.0.14
python-dateutil~=2.8.2
setuptools~=65.5.1
requests~=2.31.0
python-dotenv~=1.0.0
psycopg2~=2.9.6